+++ PHP/JSON-RPC/example_local.php
@@ -9,8 +9,8 @@ try {
     //111c1111111111cc11111111 - test group_id
     //SeCrEtToKeNvAlUe - test token
     $ssapi = new SSAPI(
-        //"127.0.0.1",
-        "api.warder.tk",
+        "127.0.0.1",
+        //"api.warder.tk",
         //"ssapi.selectspecs.com",
         8843,
         "56b32b778ec929c4110cbbfc",
@@ -39,7 +39,7 @@ function result_print($result){
 
 //Insert items
 $item4save = [];
-include(dirname(__FILE__).'/examples/item4.php');
+//include(dirname(__FILE__).'/examples/item4.php');
 //$result = $ssapi->items(NULL, $item4save, SSAPI_RETURN_RESULT);
 //echo "<h2 style='color: blueviolet;'>ITEMS INSERT:</h2>";
 //result_print($result);
@@ -50,8 +50,13 @@ include(dirname(__FILE__).'/examples/item4.php');
 //result_print($result);
 
 //Select LAST items by query
-$result = $ssapi->items_last_updated("2012-01-01 00:00:00.000Z", NULL, SSAPI_CONVERTER_WEB);
-echo "<h2 style='color: blueviolet;'>SELECT LAST:</h2>";
+//$result = $ssapi->items_last_updated("2012-01-01 00:00:00.000Z", NULL, SSAPI_CONVERTER_WEB);
+//echo "<h2 style='color: blueviolet;'>SELECT LAST:</h2>";
+//result_print($result);
+
+//Select MIN&MAX LAST items by query
+$result = $ssapi->items_last_updated_minmax("2012-01-01T00:00:00.000Z", NULL);
+echo "<h2 style='color: blueviolet;'>SELECT MIN&MAX LAST:</h2>";
 result_print($result);
 
 /*